OPL-91: Move Policy business logic from GQL mutations to services, signal for calcrule
